+++ src/Entity/SignalementEsin.php
@@ -113,6 +113,26 @@ class SignalementEsin
      */
     private $codeSiteTrois;
 
+    /**
+     * @ORM\Column(type="text", nullable=true)
+     */
+    private $investigation;
+
+    /**
+     * @ORM\Column(type="text", nullable=true)
+     */
+    private $hypotheseCause;
+
+    /**
+     * @ORM\Column(type="text", nullable=true)
+     */
+    private $justification;
+
+    /**
+     * @ORM\Column(type="text", nullable=true)
+     */
+    private $praticienHygiene;
+
     public function getId(): ?int
     {
         return $this->id;
@@ -369,6 +389,54 @@ class SignalementEsin
     public function setCodeSiteTrois(int $codeSiteTrois): self
     {
         $this->codeSiteTrois = $codeSiteTrois;
+
+        return $this;
+    }
+
+    public function getInvestigation(): ?string
+    {
+        return $this->investigation;
+    }
+
+    public function setInvestigation(?string $investigation): self
+    {
+        $this->investigation = $investigation;
+
+        return $this;
+    }
+
+    public function getHypotheseCause(): ?string
+    {
+        return $this->hypotheseCause;
+    }
+
+    public function setHypotheseCause(?string $hypotheseCause): self
+    {
+        $this->hypotheseCause = $hypotheseCause;
+
+        return $this;
+    }
+
+    public function getJustification(): ?string
+    {
+        return $this->justification;
+    }
+
+    public function setJustification(?string $justification): self
+    {
+        $this->justification = $justification;
+
+        return $this;
+    }
+
+    public function getPraticienHygiene(): ?string
+    {
+        return $this->praticienHygiene;
+    }
+
+    public function setPraticienHygiene(?string $praticienHygiene): self
+    {
+        $this->praticienHygiene = $praticienHygiene;
 
         return $this;
     }
